+++ CHANGELOG.md
@@ -15,6 +15,8 @@ the string at the given key_sequence
 https://www.chemie.uni-bonn.de/grimme/de/software/dft-d3/get_dft-d3. 2. Some
 calculators which return a 3x3 matrix for stress will break. One can modify
 ASE source for this as ASIMTools can't go into the calculator code.
+- asim_check now also reports the job_ids
+- get_atoms now allows addition of FixAtoms constraint
  
 ### Changed
 - VASP interface changed to align more with pymatgen